+++ pages/products/list/[[category]].vue
@@ -1,33 +1,17 @@
 <script setup lang="ts">
 import { ref } from 'vue'
 import { useProducts } from '~/composables/useProducts'
-import { useLocalStorage } from '@vueuse/core'
+import { useFavoritesStore } from '~/stores'
 
-const route = useRoute()
 const products = ref(useProducts())
+const favoritesStore = useFavoritesStore()
 
 definePageMeta({
     layout: 'products',
 })
 
-interface FavoriteItem {
-    id: number
-    amount: number
-    price: number
-}
-
-const favorite = useLocalStorage<FavoriteItem[]>("favorite", [])
-
-const handleFavorite = (id: number) => {
-    if (favorite.value.some((obj) => obj.id === id)) {
-        favorite.value = favorite.value.filter((el) => el.id !== id)
-    } else {
-        favorite.value.push({
-            id: id,
-            amount: 1,
-            price: 0,
-        })
-    }
+function handleFavorite(id: number) {
+    favoritesStore.toggle(id)
 }
 </script>
 
@@ -53,7 +37,7 @@ const handleFavorite = (id: number) => {
             </div>
             <div class="list-content_grid">
                 <ProductCard v-for="product in products" :key="product.id" :product="product"
-                    @favor="handleFavorite" :favored="favorite.some((obj) => obj.id === product.id)" />
+                    @favor="handleFavorite" :favored="favoritesStore.isFavorite(product.id)" />
             </div>
         </div>
     </div>