+++ stores/index.ts
@@ -1,0 +1,11 @@
+/**
+ * @file stores/index.ts
+ * @description Pinia stores barrel export
+ * @author Jenning Schaefer
+ * @license MIT
+ */
+
+export { useCartStore } from './cart'
+export { useUserStore } from './user'
+export { useFavoritesStore } from './favorites'
+export { useUiStore } from './ui'